Mapping stops in placed buildings
